+++ examples/run-examples.js
@@ -75,6 +75,30 @@ exampleFiles = [
   // 'ex-create-secret.js',
   // 'ex-get-secret.js',
   // 'ex-list-secrets.js',
+  //  'ex-send-sms.js',
+  // 'ex-create-user.js',
+  'ex-get-user.js',
+  'ex-get-user-conversations.js',
+  // 'ex-update-user.js',
+  // 'ex-delete-user.js',
+  'ex-get-users.js',
+  // 'ex-create-conversation.js',
+  // 'ex-create-member.js',
+  // 'ex-update-member.js',
+  // 'ex-delete-member.js',
+  'ex-get-members.js',
+  'ex-get-member.js',
+  // 'ex-create-event.js',
+  // 'ex-delete-event.js',
+  'ex-get-event.js',
+  'ex-get-events.js',
+  // 'ex-create-read-update-delete-conversation.js'
+  'ex-get-conversation.js',
+  // 'ex-delete-conversation.js',
+  // 'ex-update-conversation.js'
+  'ex-get-conversations.js'
+  // 'ex-stream-to-call.js',
+  // 'ex-talk-to-call.js'
   // 'ex-revoke-secret.js'
 ];