+++ test/Users-test.js
@@ -1,0 +1,124 @@
+import chai, { expect } from "chai";
+import sinon from "sinon";
+import sinonChai from "sinon-chai";
+
+chai.use(sinonChai);
+
+import ResourceTestHelper from "./ResourceTestHelper";
+
+import Users from "../lib/Users";
+import HttpClient from "../lib/HttpClient";
+import Credentials from "../lib/Credentials";
+
+var creds = Credentials.parse({
+  applicationId: "some-id",
+  privateKey: __dirname + "/private-test.key"
+});
+var emptyCallback = () => {};
+
+describe("Users", () => {
+  var httpClientStub = null;
+  var users = null;
+
+  beforeEach(() => {
+    httpClientStub = sinon.createStubInstance(HttpClient);
+    var options = {
+      httpClient: httpClientStub
+    };
+    users = new Users(creds, options);
+  });
+
+  it("should allow a user to be created", () => {
+    var params = {};
+    users.create(params, emptyCallback);
+
+    var expectedRequestArgs = ResourceTestHelper.requestArgsMatch(params, {
+      path: `${Users.PATH}`
+    });
+    expect(httpClientStub.request).to.have.been.calledWith(
+      sinon.match(expectedRequestArgs),
+      emptyCallback
+    );
+  });
+
+  it("should get a collection of Users", () => {
+    users.get({}, emptyCallback);
+
+    var expectedRequestArgs = ResourceTestHelper.requestArgsMatch(null, {
+      method: "GET",
+      body: undefined,
+      path: `${Users.PATH}`
+    });
+
+    expect(httpClientStub.request).to.have.been.calledWith(
+      sinon.match(expectedRequestArgs),
+      emptyCallback
+    );
+  });
+
+  it("should get a single user using a user ID", () => {
+    const userId = "USR-aaabbbccc-111222333";
+    users.get(userId, emptyCallback);
+
+    var expectedRequestArgs = ResourceTestHelper.requestArgsMatch(null, {
+      method: "GET",
+      body: undefined,
+      path: `${Users.PATH}/${userId}`
+    });
+
+    expect(httpClientStub.request).to.have.been.calledWith(
+      sinon.match(expectedRequestArgs),
+      emptyCallback
+    );
+  });
+
+  it("should get conversations for a user", () => {
+    const userId = "USR-aaabbbccc-111222333";
+    users.getConversations(userId, emptyCallback);
+
+    var expectedRequestArgs = ResourceTestHelper.requestArgsMatch(null, {
+      method: "GET",
+      body: undefined,
+      path: `${Users.PATH}/${userId}/conversations`
+    });
+
+    expect(httpClientStub.request).to.have.been.calledWith(
+      sinon.match(expectedRequestArgs),
+      emptyCallback
+    );
+  });
+
+  it("should allow a user to be updated", () => {
+    const userId = "USR-aaabbbccc-111222333";
+    var params = {
+      action: "join"
+    };
+    users.update(userId, params, emptyCallback);
+
+    var expectedRequestArgs = ResourceTestHelper.requestArgsMatch(params, {
+      method: "PUT",
+      path: `${Users.PATH}/${userId}`
+    });
+
+    expect(httpClientStub.request).to.have.been.calledWith(
+      sinon.match(expectedRequestArgs),
+      emptyCallback
+    );
+  });
+
+  it("should allow a user to be deleted", () => {
+    const userId = "USR-aaabbbccc-111222333";
+    users.delete(userId, emptyCallback);
+
+    var expectedRequestArgs = ResourceTestHelper.requestArgsMatch(null, {
+      method: "DELETE",
+      body: undefined,
+      path: `${Users.PATH}/${userId}`
+    });
+
+    expect(httpClientStub.request).to.have.been.calledWith(
+      sinon.match(expectedRequestArgs),
+      emptyCallback
+    );
+  });
+});